+++ i18n/ru.json
@@ -1,18 +1,6 @@
 {
-  "Header": {
-    "Projects": "Проекты",
-    "Goals": "Цели",
-    "Issues": "Задачи",
-    "Explore": "Обзор",
-    "Boards": "Доски",
-    "New goal": "Новая цель",
-    "New project": "Новый проект",
-    "Invite users": "Пригласить",
-    "Sign out": "Выйти"
-  },
   "index": {
-    "title": "Taskany Goals — Персональная подборка",
-    "Load more": "Загрузить еще"
+    "title": "Taskany"
   },
   "projects": {
     "index": {
@@ -68,6 +56,10 @@
     }
   },
   "goals": {
+    "index": {
+      "title": "Taskany Goals — Персональная подборка",
+      "Load more": "Загрузить еще"
+    },
     "new": {
       "title": "Taskany Goals — Новая цель",
       "Create new goal": "Создание новой цели",
@@ -113,6 +105,17 @@
     "item": {
       "by": ""
     }
+  },
+  "Header": {
+    "Projects": "Проекты",
+    "Goals": "Цели",
+    "Issues": "Задачи",
+    "Explore": "Обзор",
+    "Boards": "Доски",
+    "New goal": "Новая цель",
+    "New project": "Новый проект",
+    "Invite users": "Пригласить",
+    "Sign out": "Выйти"
   },
   "TagCompletion": {
     "Something went wrong 😿": "Упс! Что-то пошло не так 😿",
